+++ Runtime/Data/Message.cs
@@ -1,9 +1,0 @@
-namespace Neocortex.Data
-{
-    public struct Message
-    {
-        public string content;
-        public string sender;
-        public string createdAt;
-    }
-}